+++ lib/brasil_fields.dart
@@ -1,5 +1,6 @@
 library brasil_fields;
 
+export 'formatter/cartao_bancario_input_formatter.dart';
 export 'formatter/cep_input_formatter.dart';
 export 'formatter/cnpj_input_formatter.dart';
 export 'formatter/cpf_input_formatter.dart';
@@ -7,6 +8,8 @@ export 'formatter/data_input_formatter.dart';
 export 'formatter/real_input_formatter.dart';
 export 'formatter/telefone_input_formatter.dart';
 export 'formatter/hora_input_formatter.dart';
+export 'formatter/hora_input_formatter.dart';
+export 'formatter/validade_cartao_input_formatter.dart';
 export 'modelos/estados.dart';
 export 'modelos/meses.dart';
 export 'modelos/regioes.dart';