+++ lib/formatter/data_input_formatter.dart
@@ -8,15 +8,15 @@ class DataInputFormatter extends TextInputFormatter {
   @override
   TextEditingValue formatEditUpdate(
       TextEditingValue oldValue, TextEditingValue newValue) {
-    final int newTextLength = newValue.text.length;
-    int selectionIndex = newValue.selection.end;
+    final newTextLength = newValue.text.length;
+    var selectionIndex = newValue.selection.end;
 
     if (newTextLength > maxLength) {
       return oldValue;
     }
 
-    int usedSubstringIndex = 0;
-    final StringBuffer newText = StringBuffer();
+    var usedSubstringIndex = 0;
+    final newText = StringBuffer();
 
     if (newTextLength >= 3) {
       newText.write(newValue.text.substring(0, usedSubstringIndex = 2) + '/');