+++ lib/formatter/hora_input_formatter.dart
@@ -8,11 +8,11 @@ class HoraInputFormatter extends TextInputFormatter {
   @override
   TextEditingValue formatEditUpdate(
       TextEditingValue oldValue, TextEditingValue newValue) {
-    final int newTextLength = newValue.text.length;
-    int selectionIndex = newValue.selection.end;
+    final newTextLength = newValue.text.length;
+    var selectionIndex = newValue.selection.end;
 
-    int usedSubstringIndex = 0;
-    final StringBuffer newText = StringBuffer();
+    var usedSubstringIndex = 0;
+    final newText = StringBuffer();
 
     if (newTextLength > maxLength) {
       return oldValue;